[CIVIC-1010] Moved icon and background to base component.
