+++ docroot/themes/contrib/civictheme/.storybook/main.js
@@ -10,8 +10,8 @@ const webpack = require('webpack');
 const path = require('path');
 const custom = require('../webpack/webpack.prod');
 const scssVariables = require('./importer.scss_variables');
-const iconUtils = require('../components/01-atoms/icon/icon.utils');
-const backgroundUtils = require('../components/01-atoms/background/background.utils');
+const iconUtils = require('../components/00-base/icon/icon.utils');
+const backgroundUtils = require('../components/00-base/background/background.utils');
 const logoUtils = require('../components/02-molecules/logo/logo.utils');
 const addonConfig = require('./addon-config').default();
