+++ docroot/themes/contrib/civictheme/civictheme_library/components/00-base/background/background.stories.js
@@ -2,7 +2,7 @@
 import { select } from '@storybook/addon-knobs';
 
 export default {
-  title: 'Atoms/Background',
+  title: 'Base/Background',
   parameters: {
     layout: 'centered',
   },
