+++ docroot/themes/contrib/civictheme/civictheme_library/components/00-base/background/background.utils.js
