+++ docroot/themes/contrib/civictheme/civictheme_library/components/00-base/icon/icon.stories.js
@@ -4,7 +4,7 @@ import { radios, select, text } from '@storybook/addon-knobs';
 import CivicThemeIcon from './icon.twig';
 
 export default {
-  title: 'Atoms/Icon',
+  title: 'Base/Icon',
   parameters: {
     layout: 'centered',
   },
