+++ docroot/themes/contrib/civictheme/civictheme_library/components/00-base/icon/icon.utils.js
