+++ docroot/themes/contrib/civictheme/civictheme_library/components/01-atoms/button/button.twig
@@ -38,7 +38,7 @@
 {% set button_icon %}
   {% if icon %}
     <span class="civictheme-button__icon">
-		  {% include '@atoms/icon/icon.twig' with {
+		  {% include '@base/icon/icon.twig' with {
         symbol: icon,
         size: icon_size,
       } only %}
