+++ docroot/themes/contrib/civictheme/civictheme_library/components/01-atoms/link/link.twig
@@ -24,7 +24,7 @@
 
 {% if icon %}
   {% set link_icon %}
-    {% include '@atoms/icon/icon.twig' with {
+    {% include '@base/icon/icon.twig' with {
       symbol: icon,
       size: icon_size,
     } only %}
