+++ docroot/themes/contrib/civictheme/civictheme_library/components/01-atoms/tag/tag.twig
@@ -30,7 +30,7 @@
 
 {% set icon %}
   {% if icon %}
-    {% include '@atoms/icon/icon.twig' with {
+    {% include '@base/icon/icon.twig' with {
       symbol: icon,
       size: 'small',
     } only %}
