+++ docroot/themes/contrib/civictheme/civictheme_library/components/02-molecules/figure/figure.twig
@@ -30,7 +30,7 @@
 
     {% if caption is not empty %}
       <figcaption class="civictheme-figure__caption">
-        {% include '@atoms/icon/icon.twig' with {
+        {% include '@base/icon/icon.twig' with {
           symbol: 'multimedia_image_3',
           color: 'neutral_60',
         } %}
