+++ docroot/themes/contrib/civictheme/civictheme_library/components/02-molecules/next-steps/next-steps.twig
@@ -52,7 +52,7 @@
 
             {% block icon %}
               <div class="civictheme-next-steps__icon civictheme-next-steps__read-more">
-                {% include '@atoms/icon/icon.twig' with {
+                {% include '@base/icon/icon.twig' with {
                   symbol: 'upper-right-arrow',
                   color: 'primary'
                 } only %}
