+++ docroot/themes/contrib/civictheme/civictheme_library/components/02-molecules/subject-card/subject-card.twig
@@ -50,13 +50,13 @@
             {% if url is not empty %}
               <span>
                 {% if is_external %}
-                  {% include '@atoms/icon/icon.twig' with {
+                  {% include '@base/icon/icon.twig' with {
                     symbol: 'upper-right-arrow',
                     color: 'primary',
                     modifier_class: 'civictheme-subject-card__icon--external'
                   } only %}
                 {% else %}
-                  {% include '@atoms/icon/icon.twig' with {
+                  {% include '@base/icon/icon.twig' with {
                     symbol: 'upper-right-arrow',
                     color: 'primary',
                     modifier_class: 'civictheme-subject-card__icon--arrow',
