+++ docroot/themes/contrib/civictheme/civictheme_library/components/03-organisms/accordion/accordion.twig
@@ -43,7 +43,7 @@
                       {% if fold.title is not empty %}
                         <span>{{ fold.title }}</span>
                       {% endif %}
-                      {% include '@atoms/icon/icon.twig' with {
+                      {% include '@base/icon/icon.twig' with {
                         symbol: 'arrows_downarrow_1'
                       } only %}
                     </button>
