+++ docroot/themes/contrib/civictheme/civictheme_library/components/03-organisms/breadcrumb/breadcrumb.twig
@@ -40,7 +40,7 @@
                 attributes: loop.index == links_count ? 'aria-current="location"' : '',
               } only %}
               {% if (loop.index < links_count) %}
-                {% include '@atoms/icon/icon.twig' with {
+                {% include '@base/icon/icon.twig' with {
                   symbol: 'arrows_rightarrow',
                 } only %}
               {% endif %}
@@ -52,7 +52,7 @@
 
       {% if parent is not empty %}
         <li class="civictheme-breadcrumb__links--link mobile-only">
-          {% include '@atoms/icon/icon.twig' with {
+          {% include '@base/icon/icon.twig' with {
             symbol: 'arrows_leftarrow_2',
           } only %}
           {% include '@atoms/link/link.twig' with {
