+++ docroot/themes/contrib/civictheme/civictheme_library/components/03-organisms/large-filter/large-filter.twig
@@ -110,7 +110,7 @@
               <div class="civictheme-large-filter__clear-container" data-large-filter-clear-container>
                 <button class="civictheme-large-filter__clear-all" data-large-filter-clear>
                   <span class="civictheme-large-filter__clear-all-label">{{ clear_text }}</span>
-                  {% include '@atoms/icon/icon.twig' with {
+                  {% include '@base/icon/icon.twig' with {
                     symbol: 'userinterface_close_2',
                     color: 'error',
                   } only %}
