+++ docroot/themes/contrib/civictheme/civictheme_library/components/03-organisms/message/message.twig
@@ -26,7 +26,7 @@
 <div class="civictheme-message messages {{ modifier_class }}" role="{% if type == 'error' %}error{% else %}contentinfo{% endif %}" aria-label="{{ type }}" {% if type == 'error' %}aria-live="polite"{% endif %}>
   {% if icons[type] is defined %}
     <div class="civictheme-message__icon">
-      {% include "@atoms/icon/icon.twig" with {
+      {% include "@base/icon/icon.twig" with {
         symbol: icons[type],
         modifier_class: "%s civictheme-icon--%s"|format(theme_class, type)
       } only %}
